Improved documentation on QoS requirements objects. Fixed remaining uses of TimeSpan.Milliseconds to use .TotalMilliseconds Fixed FormatExceptions when DebugUtils.Verbose = true

+++ GTUnitTests/GTClientServerTests.cs
@@ -110,7 +110,7 @@ namespace GT.UnitTests
         public bool ErrorOccurred { get { return errorOccurred; } }
 
         public int ServerSleepTime {
-            get { return config.TickInterval.TotalMilliseconds; }
+            get { return (int)config.TickInterval.TotalMilliseconds; }
             set { config.TickInterval = TimeSpan.FromMilliseconds(value); }
         }
